+++ src/AssetClass/Receivable.hs
@@ -105,6 +105,7 @@ instance Asset Receivable where
   splitWith r@(Invoice (ReceivableInfo sd ob oa dd ft) st) rs 
     = [ (Invoice (ReceivableInfo sd (mulBR ob ratio) (mulBR oa ratio) dd ft) st) | ratio <- rs ]
 
+  -- Defaulted Invoice
   projCashflow r@(Invoice (ReceivableInfo sd ob oa dd ft) (Defaulted _))
                asOfDay
                massump@(A.ReceivableAssump _ amr ams, _ , _)
@@ -117,6 +118,7 @@ instance Asset Receivable where
       (futureTxns,historyM)= CF.cutoffTrs asOfDay (patchLossRecovery (initTxn++txns) amr)
 
 
+  -- Performing Invoice
   projCashflow r@(Invoice (ReceivableInfo sd ob oa dd ft) Current) 
                asOfDay
                massump@(A.ReceivableAssump (Just A.DefaultAtEnd) amr ams, _ , _)
@@ -126,8 +128,10 @@ instance Asset Receivable where
       payDate = dd
       feeDue = calcDueFactorFee r payDate
       initTxn = [CF.ReceivableFlow sd ob 0 0 0 0 0 0 Nothing]
-      
-      txns = [CF.ReceivableFlow payDate 0 0 0 0 ob 0 ob Nothing]
+      realizedLoss = case amr of
+                      Nothing -> ob
+                      Just _ -> 0
+      txns = [CF.ReceivableFlow payDate 0 0 0 0 ob 0 realizedLoss Nothing]
       (futureTxns,historyM)= CF.cutoffTrs asOfDay $ txns++(buildRecoveryCfs payDate ob amr)
 
   projCashflow r@(Invoice (ReceivableInfo sd ob oa dd ft) Current) 
@@ -148,5 +152,9 @@ instance Asset Receivable where
       feePaid = min afterDefaultBal afterDefaultFee
       principal = max 0 $ afterDefaultBal - feePaid
       
-      txns = [CF.ReceivableFlow payDate 0 0 principal feePaid defaultAmt 0 defaultAmt Nothing]
+      realizedLoss = case amr of
+                      Nothing -> defaultAmt
+                      Just _ -> 0
+      
+      txns = [CF.ReceivableFlow payDate 0 0 principal feePaid defaultAmt 0 realizedLoss Nothing]
       (futureTxns,historyM) = CF.cutoffTrs asOfDay $ txns++(buildRecoveryCfs payDate defaultAmt amr) `debug` ("recovery flow"++ show (buildRecoveryCfs payDate defaultAmt amr))